Week 3 preso complete
API + MySQL Connection examples done.

The slides here work a lot with the examples.
If the class has a lot of questions, I might fill
out the slides a little more.

There's a lot to get through here, so we may have
to touch on the db connectivity next week.